refactor: update directory name from components -> app
